pyproject.toml
[project]
name = "ai-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-anthropic>=1.3.4",
    "langchain-core>=1.2.16",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
]
